🔒 Warden: [security fix] Fix data race in test harness environment variable mutation
🦠 Threat: `unsafe { std::env::set_var }` was being called by three different inline `EnvGuard` instances across the codebase (`autumn/src/app.rs`, `autumn/src/middleware/dev.rs`, `autumn-cli/src/dev.rs`), each using their own local module `OnceLock`. Because Cargo runs test binaries concurrently, this allowed multi-threaded parallel mutation of the process-global environment, leading to Undefined Behavior (data races in libc environment structures) and potential test flakes or segmentation faults.
🛡️ Defense: Centralized the `EnvGuard` struct into a single `test_utils` module per crate. This ensures that the `OnceLock<Mutex<()>>` correctly serializes all environment mutations during `cargo test` runs, preventing any concurrent access.
💥 Severity: Medium - Undefined Behavior in the test harness could lead to flaky CI pipelines and arbitrary memory corruption during test execution.
🧪 Verification: Ran `cargo check`, `cargo test -p autumn-web --tests`, `cargo test -p autumn-cli --tests`, and `cargo clippy --all-targets --all-features`. Verified that tests still pass, no timeouts occur, and the `EnvGuard` cleanly synchronizes.
