+++ autumn-cli/src/main.rs
@@ -7,6 +7,8 @@ mod migrate;
 mod monitor;
 mod new;
 mod setup;
+#[cfg(test)]
+pub(crate) mod test_utils;
 
 /// The Autumn web framework CLI.
 #[derive(Parser)]
