+++ autumn/src/lib.rs
@@ -83,6 +83,8 @@ pub mod hooks;
 #[cfg(feature = "db")]
 pub mod migrate;
 pub mod router;
+#[cfg(test)]
+pub(crate) mod test_utils;
 #[cfg(feature = "db")]
 pub use hooks::{
     DraftField, FieldDiff, MutationContext, MutationHooks, MutationOp, NoHooks, Patch, UpdateDraft,
